Encrypt moved to ApiHelper
