+++ OngProject/Core/Helper/JwtHelper.cs
@@ -20,16 +20,6 @@ namespace OngProject.Core.Helper
         {
             _configuration = configuration;
         }
-        public static string Encrypt(string str)
-        {
-            SHA256 sha256 = SHA256Managed.Create();
-            ASCIIEncoding encoding = new ASCIIEncoding();
-            byte[] stream = null;
-            StringBuilder sb = new StringBuilder();
-            stream = sha256.ComputeHash(encoding.GetBytes(str));
-            for (int i = 0; i < stream.Length; i++) sb.AppendFormat("{0:x2}", stream[i]);
-            return sb.ToString();
-        }
 
         public string GetToken(Users user)
         {
